refactor: allow loading of individual replay files for stats
